types(server): split in files
