+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-action-common-v2.ts
@@ -1,0 +1,28 @@
+import { WorkspaceMigrationFieldActionV2 } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-field-action-v2';
+import { WorkspaceMigrationV2ObjectAction } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-object-action-v2';
+import { WorkspaceMigrationUniquenessActionV2 } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-uniqueness-action-v2';
+import { WorkspaceMigrationIndexActionV2 } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migratoin-index-action-v2';
+import { WorkspaceMigrationRelationActionV2 } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migratoin-relation-action-v2';
+
+type UniqueIdentifierRecord<TTarget extends string> = {
+  [P in `${TTarget}UniqueIdentifier`]: string;
+};
+
+export type ObjectMetadataUniqueIdentifier =
+  UniqueIdentifierRecord<'objectMetadata'>;
+export type FieldMetadataUniqueIdentifier =
+  UniqueIdentifierRecord<'fieldMetadata'>;
+
+export type FromTo<T> = {
+  from: T;
+  to: T;
+};
+
+export type WorkspaceMigrationActionV2 =
+  | WorkspaceMigrationRelationActionV2
+  | WorkspaceMigrationV2ObjectAction
+  | WorkspaceMigrationFieldActionV2
+  | WorkspaceMigrationUniquenessActionV2
+  | WorkspaceMigrationIndexActionV2;
+
+export type WorkspaceMigrationActionTypeV2 = WorkspaceMigrationActionV2['type'];
