+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-v2.ts
@@ -1,4 +1,4 @@
-import { WorkspaceMigrationActionV2 } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-action-v2';
+import { WorkspaceMigrationActionV2 } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-action-common-v2';
 
 export interface WorkspaceMigrationV2<
   TActions extends WorkspaceMigrationActionV2 = WorkspaceMigrationActionV2,
