+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/types/workspace-migratoin-index-action-v2.ts
@@ -1,0 +1,11 @@
+export interface CreateIndexAction {
+  type: 'create_index';
+}
+
+export interface DeleteIndexAction {
+  type: 'delete_index';
+}
+
+export type WorkspaceMigrationIndexActionV2 =
+  | CreateIndexAction
+  | DeleteIndexAction;
