+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/types/workspace-migratoin-relation-action-v2.ts
@@ -1,0 +1,16 @@
+export interface CreateRelationAction {
+  type: 'create_relation';
+}
+
+export interface UpdateRelationAction {
+  type: 'update_relation';
+}
+
+export interface DeleteRelationAction {
+  type: 'delete_relation';
+}
+
+export type WorkspaceMigrationRelationActionV2 =
+  | CreateRelationAction
+  | UpdateRelationAction
+  | DeleteRelationAction;
